vector/requirements.txt
flask
chromadb
python-dotenv
